chore: address more comments
